requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-autoslug==1.9.8
django-social-auth==0.7.28
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
oauth2==1.9.0.post1
oauthlib==3.2.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-openid==2.2.5
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
social==0.1.dev3
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
waitress==2.1.2
whitenoise==6.2.0
